requirements.txt
# Core Framework
agno>=0.1.0
phidata>=2.0.0  # Fallback if agno is not available

# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1

# AI & LLM
openai>=1.12.0
anthropic>=0.18.0

# Environment & Configuration
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# HTTP & API Integration
httpx>=0.26.0
requests>=2.31.0

# Audio Processing (for DET speech tasks)
speechrecognition>=3.10.1
pydub>=0.25.1

# Utilities
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pendulum>=3.0.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.4
black>=24.1.1
ruff>=0.2.1
